[refactoring] group all 'OriginItem' classes togther
Group together all the classes relating to the OriginItem, renaming them
in the process (Position is a bit too ambiguous, especially as some of
the items aren't directly positions, but things like constituencies or
legislative terms)
